almost finishing Le

+++ srcs/ft_printf/parse.c
@@ -67,7 +67,7 @@ char	*parse_efg(t_sub *sub, va_list ap)
 	else if (TYPE & 0x80C00000)
 		S = ft_ftosn(va_arg(ap, double), PREC);
 	else if (TYPE == 0x40C00000)
-		S = ft_ftosn(va_arg(ap, long double), PREC);
+		S = ft_ftosn((double)va_arg(ap, long double), PREC);
 	return ((S) ? crop_efg(sub) : S);
 }
